requirements.txt
ollama
python-dotenv
rapidfuzz
discord.py
pygame
python-statemachine
elevenlabs
websocket
asyncio
transformers
tensorflow
tf-keras